feat(cv): add direct links to English and French versions
Added direct links to the English and French versions of the resume in the CV page. This allows users to easily access the specific version they need without having to navigate through the entire page. The links are labeled "Direct link English" and "Direct link French" respectively. The links point to the corresponding PDF files in the assets folder.

The motivation for this change is to improve the user experience and make it more convenient for users to access the specific version of the resume they are interested in.